--- packages/core-mobile/app/new/common/hooks/useIsOnline.ts
@@ -1,19 +1,0 @@
-import { useEffect, useState } from 'react'
-import NetInfo from '@react-native-community/netinfo'
-
-export const useIsOnline = (): boolean => {
-  const [isOnline, setIsOnline] = useState(false)
-
-  // manage online status
-  useEffect(() => {
-    return NetInfo.addEventListener(state => {
-      const online =
-        state.isConnected != null &&
-        state.isConnected &&
-        Boolean(state.isInternetReachable)
-      setIsOnline(online)
-    })
-  }, [])
-
-  return isOnline
-}
